==> R/05_xlsx_3yrs_Scot.R <==

# TO DO -----
# update publication date in Readme

# prelims ----------------------------------------------------------------------

library(tidyverse)
library(openxlsx)

source("R/00_functions.R")
source("R/00_strings.R")

filename <- "website/xls/data2023.xlsx"
mytables <- readRDS("data/tables.rds")

# tables not to be included in spreadsheet
mytables <- within(mytables,
                   rm(corign_ad, distribution, median_ages, equivalence))

latestyear <- max(levels(labels$years$formatted))
latestperiod <- max(levels(labels$years$periods))

# headers ----------------------------------------------------------------------

myheaders <- list(

  # * -- headline measures -----------------------------------------------------

  list(header = c("Relative poverty after housing costs",
                  "This worksheet contains four tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Rate: Proportion of people in each group who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition: Proportion of people in relative poverty (below 60% of UK median income after housing costs) who are in each group, Scotland",
                  "Number: Number of people in each group who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families in each group in the combined three-year survey sample, Scotland")),
  list(header = c("Absolute poverty after housing costs",
                  "This worksheet contains four tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Rate: Proportion of people in each group who are in absolute poverty (below 60% of the 2010/11 UK median income after housing costs), Scotland",
                  "Composition: Composition of those in severe poverty: Proportion of people in severe poverty (below 60% of the 2010/11 UK median income after housing costs) who are in each group, Scotland",
                  "Number: Number of people in each group who are in severe poverty (below 60% of the 2010/11 UK median income after housing costs), Scotland",
                  "Sample size: Number of families in each group in the combined three-year survey sample, Scotland")),
  list(header = c("Severe poverty after housing costs",
                  "This worksheet contains four tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Rate: Proportion of people in each group who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition: Composition of those in severe poverty: Proportion of people in severe poverty (below 50% of UK median income after housing costs) who are in each group, Scotland",
                  "Number: Number of people in each group who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families in each group in the combined three-year survey sample, Scotland")),
  list(header = c("Relative poverty before housing costs",
                  "This worksheet contains four tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Rate: Relative poverty rate: Proportion of people in each group who are in relative poverty (below 60% of UK median income before housing costs), Scotland",
                  "Composition: Composition of those in relative poverty: Proportion of people in relative poverty (below 60% of UK median income before housing costs) who are in each group, Scotland",
                  "Number: Relative poverty numbers: Number of people in each group who are in relative poverty (below 60% of UK median income before housing costs), Scotland",
                  "Sample size: Number of families in each group in the combined three-year survey sample, Scotland")),
  list(header = c("Absolute poverty before housing costs",
                  "This worksheet contains four tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Rate: Proportion of people in each group who are in absolute poverty (below 60% of the 2010/11 UK median income before housing costs), Scotland",
                  "Composition: Composition of those in severe poverty: Proportion of people in severe poverty (below 60% of the 2010/11 UK median income before housing costs) who are in each group, Scotland",
                  "Number: Number of people in each group who are in severe poverty (below 60% of the 2010/11 UK median income before housing costs), Scotland",
                  "Sample size: Number of families in each group in the combined three-year survey sample, Scotland")),
  list(header = c("Severe poverty before housing costs",
                  "This worksheet contains four tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Rate: Proportion of people in each group who are in severe poverty (below 50% of UK median income before housing costs), Scotland",
                  "Composition: Composition of those in severe poverty: Proportion of people in severe poverty (below 50% of UK median income before housing costs) who are in each group, Scotland",
                  "Number: Number of people in each group who are in severe poverty (below 50% of UK median income before housing costs), Scotland",
                  "Sample size: Number of families in each group in the combined three-year survey sample, Scotland")),
  list(header = c("Children's combined low income and material deprivation",
                  "This worksheet contains three tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: The latest two estimates cannot be directly compared to previous estimates. This is because the latest two periods cover a time when families were less able to undertake certain activities due to the pandemic, and not necessarily because they couldn't afford to. This changed how people responded to the material deprivation questions.",
                  "Note: The definition of material deprivation changed in 2010/11, creating a break in the time series.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Rate: Proportion of children who are in combined low income (below 70% of UK median income) and material deprivation, Scotland",
                  "Number: Number of children who are in combined low income (below 70% of UK median income) and material deprivation, Scotland",
                  "Sample size: Sample size: Number of families with children in the combined three-year survey sample, Scotland")),
  list(header = c("Pensioners' material deprivation",
                  "This worksheet contains three tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: The latest two estimates cannot be directly compared to previous estimates. This is because the latest two periods cover a time when families were less able to undertake certain activities due to the pandemic, and not necessarily because they couldn't afford to. This changed how people responded to the material deprivation questions.",
                  "Note: Pensioner material deprivation is calculated for pensioners aged 65 or over. Pensioner material deprivation is different to other measures of poverty, including the child low income and material deprivation measure in that it is not associated with an income threshold. It captures issues such as whether poor health, disability and social isolation prevent access to goods and services, rather than solely low income.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Rate: Proportion of pensioners aged 65 and over who have limited access to goods and services, Scotland",
                  "Number: Number of pensioners aged 65 and over who have limited access to goods and services, Scotland",
                  "Sample size: Number of families with pensioners aged 65 and over in the combined three-year survey sample, Scotland")),

  # * -- people -----------------------------------------------

  list(header = c("Age of household head",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of people in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of people in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of people in each category in the combined three-year survey sample, Scotland")),

  list(header = c("Disabled household members",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: The way in which information on disabled people is collected changed several times during this timeseries. This causes breaks in the timeseries between 2001/02 and 2002/03, between 2003/04 and 2004/05, and between 2011/12 and 2012/13. Since 2012/13, disabled people are identified as those who report any physical or mental health condition(s) or illness(es) that last or are expected to last 12 months or more, and which limit their ability to carry out day-to-day activities. Therefore, care needs to be taken when considering long-term trends.",
                  "Note: The last break in the methodology caused a large change in the size of the disabled population. Therefore, the estimated numbers in poverty before and after the break cannot be directly compared and no three-year averaged data is available during the break.",
                  "Note: Data on disabled children is available from 1995/96.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of people in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of people in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families in each category in the combined three-year survey sample, Scotland")),
  list(header = c("Disabled household members, disability benefits removed from household income",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: Income from Disability Living Allowance, Attendance Allowance, and Personal Independence Payments have been excluded from income in this worksheet. This income is related to additional living costs associated with a disability.",
                  "Note: The way in which information on disabled people is collected changed several times during this timeseries. This causes breaks in the timeseries between 2001/02 and 2002/03, between 2003/04 and 2004/05, and between 2011/12 and 2012/13. Since 2012/13, disabled people are identified as those who report any physical or mental health condition(s) or illness(es) that last or are expected to last 12 months or more, and which limit their ability to carry out day-to-day activities. Therefore, care needs to be taken when considering long-term trends.",
                  "Note: The last break in the methodology caused a large change in the size of the disabled population. Therefore, the estimated numbers in poverty before and after the break cannot be directly compared and no three-year averaged data is available during the break.",
                  "Note: Data on disabled children is available from 1995/96.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of people in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of people in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families in each category in the combined three-year survey sample, Scotland")),

  list(header = c("Ethnicity - detailed breakdown (5-year average)",
                  "This worksheet contains eight tables.",
                  "Important: The latest two estimates are four-year averages. See README for more information.",
                  "Note: Ethnicity data relates to all people in a household and is based on the ethnicity of the adult with the highest income.",
                  "Note: Different ethnic groups have been combined into one for this analysis, as sample sizes are too small to reliably report on individual groups.",
                  "Note: Ethnicity data according to the categories below has been available since 2001/02.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of people in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of people in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Age: Median age of household head in each category, Scotland",
                  "Sample size: Number of families in each category in the combined five-year survey sample, Scotland")),

    list(header = c("Number of children in the household",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of people in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of people in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families in each category in the combined three-year survey sample, Scotland")),

  list(header = c("Housing tenure",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: Information on housing tenure is not available prior to 2003.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of people in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of people in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families in each category in the combined three-year survey sample, Scotland")),
  list(header = c("Urban and rural areas",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: Information on urban/rural class is not available prior to 2006.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of people in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of people in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families in each category in the combined three-year survey sample, Scotland")),

  list(header = c("Scottish Index of Multiple Deprivation areas",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of people in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of people in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of people in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of people in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families in each category in the combined three-year survey sample, Scotland")),


  # * -- children -----------------------------------------

  list(header = c("Lone parenthood",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of children in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of children in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families with children in each category in the combined three-year survey sample, Scotland")),
  list(header = c("Number of children in the household",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of children in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of children in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families with children in each category in the combined three-year survey sample, Scotland")),
  list(header = c("Child age",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of children in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of children in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families with children in each category in the combined three-year survey sample, Scotland")),
  list(header = c("Age of youngest child in the household",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: Individual child age data is not available between 2003 and 2007.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of children in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of children in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families with children in each category in the combined three-year survey sample, Scotland")),
  list(header = c("Age of mother",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: Individual age data is available from 1997, but not between 2003 and 2007.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of children in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of children in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families with children in each category in the combined three-year survey sample, Scotland")),
  list(header = c("Family economic status",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: The term 'family' here refers to the core family in a household, consisting of one or two adults and their dependent children if any. A household can contain more than one family.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of children in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of children in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families with children in each category in the combined three-year survey sample, Scotland")),
  list(header = c("Household work status",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: Information on economic status is not available prior to 1996.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of children in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of children in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families with children in each category in the combined three-year survey sample, Scotland")),
  list(header = c("Housing tenure",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: Information on housing tenure is not available prior to 2003.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of children in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of children in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families with children in each category in the combined three-year survey sample, Scotland")),
  list(header = c("Urban and rural areas",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: Information on urban/rural class is not available prior to 2006.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of children in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of children in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families with children in each category in the combined three-year survey sample, Scotland")),
  list(header = c("Disabled household members",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: The way in which information on disabled people is collected changed several times during this timeseries. This causes breaks in the timeseries between 2001/02 and 2002/03, between 2003/04 and 2004/05, and between 2011/12 and 2012/13. Since 2012/13, disabled people are identified as those who report any physical or mental health condition(s) or illness(es) that last or are expected to last 12 months or more, and which limit their ability to carry out day-to-day activities. Therefore, care needs to be taken when considering long-term trends.",
                  "Note: The last break in the methodology caused a large change in the size of the disabled population. Therefore, the estimated numbers in poverty before and after the break cannot be directly compared and no three-year averaged data is available during the break.",
                  "Note: Data on disabled children is available from 1995/96.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of children in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of children in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families with children in each category in the combined three-year survey sample, Scotland")),
  list(header = c("Disabled household members, disability benefits removed from household income",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: Income from Disability Living Allowance, Attendance Allowance, and Personal Independence Payments have been excluded from income in this worksheet. This income is related to additional living costs associated with a disability.",
                  "Note: The way in which information on disabled people is collected changed several times during this timeseries. This causes breaks in the timeseries between 2001/02 and 2002/03, between 2003/04 and 2004/05, and between 2011/12 and 2012/13. Since 2012/13, disabled people are identified as those who report any physical or mental health condition(s) or illness(es) that last or are expected to last 12 months or more, and which limit their ability to carry out day-to-day activities. Therefore, care needs to be taken when considering long-term trends.",
                  "Note: The last break in the methodology caused a large change in the size of the disabled population. Therefore, the estimated numbers in poverty before and after the break cannot be directly compared and no three-year averaged data is available during the break.",
                  "Note: Data on disabled children is available from 1995/96.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of children in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of children in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families with children in each category in the combined three-year survey sample, Scotland")),
  list(header = c("Ethnicity - detailed breakdown (5-year average)",
                  "This worksheet contains eight tables.",
                  "Important: The latest two estimates are four-year averages. See README for more information.",
                  "Note: Ethnicity data relates to all people in a household and is based on the ethnicity of the adult with the highest income.",
                  "Note: Different ethnic groups have been combined into one for this analysis, as sample sizes are too small to reliably report on individual groups.",
                  "Note: Ethnicity data according to the categories below has been available since 2001/02.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of children in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of children in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Age: Median age of family head (families with children) in each category, Scotland",
                  "Sample size: Number of families with children in each category in the combined five-year survey sample, Scotland")),
  list(header = c("Ethnicity (3-year average)",
                  "This worksheet contains eight tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: Ethnicity data relates to all people in a household and is based on the ethnicity of the adult with the highest income.",
                  "Note: Different ethnic groups have been combined into one for this analysis, as sample sizes are too small to reliably report on individual groups.",
                  "Note: Categories have changed. Previously, we combined white ethnic minorities with white British, while in this worksheet, they are combined with other ethnic minorities.",
                  "Note: Ethnicity data according to the categories above has been available since 2001/02.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of children in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of children in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of children in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Age: Median age of family head (families with children) in each category, Scotland",
                  "Sample size: Number of families with children in each category in the combined three-year survey sample, Scotland")),

  list(header = c("Priority characteristics",
                  "This worksheet contains three tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c(paste0("Relative poverty rate: Proportion of children in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland ", latestperiod),
                  paste0("Absolute poverty rate: Proportion of children in each category who are in absolute poverty (below 60% of the 2010/11 UK median income after housing costs), Scotland ", latestperiod),
                  paste0("Child material deprivation: Proportion of children in each category who are in combined low income after housing costs (below 70% of UK median income) and material deprivation, Scotland ", latestperiod),
                  paste0("Sample size: Number of families with children in each category in the combined three-year survey sample, Scotland ", latestperiod))),


  # * -- working-age adults ----------------------------------------------------

  list(header = c("Family economic status",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: The term 'family' here refers to the core family in a household, consisting of one or two adults and their dependent children if any.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of working-age adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of working-age adults in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of working-age adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of working-age adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of working-age adults in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of working-age adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families with working-age adults in each category in the combined three-year survey sample, Scotland")),
  list(header = c("Household work status",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of working-age adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of working-age adults in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of working-age adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of working-age adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of working-age adults in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of working-age adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families with working-age adults in each category in the combined three-year survey sample, Scotland")),

  # * -- adults -----------------------------------------------

  list(header = c("Family type",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: 'Pensioner couples' include working-age adults who are in a couple with a pensioner.",
                  "Note: The term 'family' here refers to the core family in a household, consisting of one or two adults and their dependent children if any. A household may contain more than one family.",
                  "Note: 'Single' adults in this analysis refer to single-adult families, not single-adult households. In some cases, single adult families may share a household with other families. This differs from the analysis in the 'Gender' worksheet, where single adults are those who share the household with no other adults.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of adults in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of adults in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of families in each category in the combined three-year survey sample, Scotland")),

  list(header = c("Adult sex",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of adults in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of adults in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of adults in each category in the combined three-year survey sample, Scotland")),

  list(header = c("Sex of single adults",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: The term 'single adult' here refers to adults who are sharing a household with no other adults. This differs from the analysis in the 'Family type' worksheet, where single adults may share the household with other families.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of single adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of single adults in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of single adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of single adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of single adults in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of single adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of single adults in each category in the combined three-year survey sample, Scotland")),

  list(header = c("Adult sexual orientation",
                  "This worksheet contains eight tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: Different groups have been combined into one for this analysis, as sample sizes are too small to reliably report on individual groups. The 'Other' group contains adults who reported to be 'Gay / lesbian', 'Bisexual' or 'Other'.",
                  "Note: Many respondents did not report their sexual orientation. This group is fairly large and therefore included as its own category.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of adults in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of adults in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Age: Median adult age in each category, Scotland",
                  "Sample size: Number of adults in each category in the combined three-year survey sample, Scotland")),

  list(header = c("Marital status",
                  "This worksheet contains seven tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Note: 'Single' refers to adults who have never been married or in a civil partnership, and are not living with a partner. 'Separated' refers to adults who are married or in a civil partnership, but are not living together because of estrangement. 'Married / Civil Partnership' includes couples who are temporarily living apart (e.g. due to serving in the armed forces). 'Adults' includes working-age adults and pensioners.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of adults in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of adults in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Sample size: Number of adults in each category in the combined three-year survey sample, Scotland")),

  list(header = c("Religion of adults - detailed breakdown (5-year average)",
                  "This worksheet contains eight tables.",
                  "Important: The latest two estimates are four-year averages. See README for more information.",
                  "Note: In these tables, 'adults' include working-age adults as well as pensioners.",
                  "Note: Different religious groups have been combined into one for this analysis, as sample sizes are too small to reliably report on individual groups.",
                  "Note: Religion data is available from 2011/12.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Relative poverty rate: Proportion of adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Composition of those in relative poverty: Proportion of adults in relative poverty (below 60% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in relative poverty: Number of adults in each category who are in relative poverty (below 60% of UK median income after housing costs), Scotland",
                  "Severe poverty rate: Proportion of adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Composition of those in severe poverty: Proportion of adults in severe poverty (below 50% of UK median income after housing costs) who are in each category, Scotland",
                  "Numbers in severe poverty: Number of adults in each category who are in severe poverty (below 50% of UK median income after housing costs), Scotland",
                  "Age: Median adult age in each category, Scotland",
                  "Sample size: Number of adults in each category in the combined five-year survey sample, Scotland")),

  # * -- food security ---------------------------------------------------------

  list(header = c("Household food security levels",
                  "This worksheet contains three tables.",
                  "Note: Shared households such as such as a group of students, or other unrelated adults were excluded from the analysis.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Proportion of people in each household food security category, Scotland 2019/20-2021/22",
                  "Number of people in each household food security category, Scotland 2019/20-2021/22",
                  "Number of families in each group in the combined two-year survey sample, Scotland 2019/20-2021/22")),

  list(header = c("Household food security by poverty status",
                  "This worksheet contains six tables.",
                  "Note: Shared households such as such as a group of students, or other unrelated adults were excluded from the analysis.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Proportion of people in relative poverty after housing costs in each household food security category, Scotland 2019/20-2021/22",
                  "Number of people in relative poverty after housing costs in each household food security category, Scotland 2019/20-2021/22",
                  "Number of families in each group in the combined two-year survey sample, Scotland 2019/20-2021/22",
                  "Proportion of people in severe poverty after housing costs in each household food security category, Scotland 2019/20-2021/22",
                  "Number of people in severe poverty after housing costs in each household food security category, Scotland 2019/20-2021/22",
                  "Number of families in each group in the combined two-year survey sample, Scotland 2019/20-2021/22")),

  list(header = c("Household food security by age",
                  "This worksheet contains three tables.",
                  "Note: Age refers to the age of the household head",
                  "Note: Shared households such as such as a group of students, or other unrelated adults were excluded from the analysis.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Proportion of people in each household food security category, Scotland 2019/20-2021/22",
                  "Number of people in each household food security category, Scotland 2019/20-2021/22",
                  "Number of families in each group in the combined two-year survey sample, Scotland 2019/20-2021/22")),

  list(header = c("Household food security by sex",
                  "This worksheet contains three tables.",
                  "Note: Sex refers to the sex of the household head",
                  "Note: Shared households such as such as a group of students, or other unrelated adults were excluded from the analysis.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Proportion of people in each household food security category, Scotland 2019/20-2021/22",
                  "Number of people in each household food security category, Scotland 2019/20-2021/22",
                  "Number of families in each group in the combined two-year survey sample, Scotland 2019/20-2021/22")),

  list(header = c("Household food security by sex (single adult households)",
                  "This worksheet contains three tables.",
                  "Note: The term 'single adult' here refers to adults who are sharing a household with no other adults. This differs from the analysis in the 'Family type' tables, where single adults may share the household with other families.",
                  "Note: Shared households such as such as a group of students, or other unrelated adults were excluded from the analysis.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Proportion of single adults in each household food security category, Scotland 2019/20-2021/22",
                  "Number of single adults in each household food security category, Scotland 2019/20-2021/22",
                  "Number of single-adult families in each group in the combined two-year survey sample, Scotland 2019/20-2021/22")),

  list(header = c("Household food security by sexual orientation",
                  "This worksheet contains three tables.",
                  "Note: Different groups have been combined into one for this analysis, as sample sizes are too small to reliably report on individual groups. The 'Other' group contains adults who reported to be 'Gay / lesbian', 'Bisexual' or 'Other'.",
                  "Note: Many respondents did not report their sexual orientation. This group is fairly large and therefore included as its own category.",
                  "Note: Shared households such as such as a group of students, or other unrelated adults were excluded from the analysis.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Proportion of adults in each household food security category, Scotland 2019/20-2021/22",
                  "Number of adults in each household food security category, Scotland 2019/20-2021/22",
                  "Number of adults in each group in the combined two-year survey sample, Scotland 2019/20-2021/22")),

  list(header = c("Household food security by disability",
                  "This worksheet contains three tables.",
                  "Note: Shared households such as such as a group of students, or other unrelated adults were excluded from the analysis.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Proportion of people in each household food security category, Scotland 2019/20-2021/22",
                  "Number of people in each household food security category, Scotland 2019/20-2021/22",
                  "Number of families in each group in the combined two-year survey sample, Scotland 2019/20-2021/22")),

  list(header = c("Household food security by ethnicity",
                  "This worksheet contains three tables.",
                  "Note: Ethnicity data relates to all people in a household and is based on the ethnicity of the adult with the highest income.",
                  "Note: Different ethnic groups have been combined into one for this analysis, as sample sizes are too small to reliably report on individual groups.",
                  "Note: Ethnicity data according to the categories below has been available since 2001/02.",
                  "Note: Shared households such as such as a group of students, or other unrelated adults were excluded from the analysis.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Proportion of people in each household food security category, Scotland 2019/20-2021/22",
                  "Number of people in each household food security category, Scotland 2019/20-2021/22",
                  "Number of families in each group in the combined two-year survey sample, Scotland 2019/20-2021/22")),

  list(header = c("Household food security by religion",
                  "This worksheet contains three tables.",
                  "Note: Different religious groups have been combined into one for this analysis, as sample sizes are too small to reliably report on individual groups.",
                  "Note: Shared households such as such as a group of students, or other unrelated adults were excluded from the analysis.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Proportion of adults in each household food security category, Scotland 2019/20-2021/22",
                  "Number of adults in each household food security category, Scotland 2019/20-2021/22",
                  "Number of adults in each group in the combined two-year survey sample, Scotland 2019/20-2021/22")),

  list(header = c("Household food security by family type",
                  "This worksheet contains three tables.",
                  "Note: 'Pensioner couples' include working-age adults who are in a couple with a pensioner.",
                  "Note: The term 'family' here refers to the core family in a household, consisting of one or two adults and their dependent children if any. A household may contain more than one family.",
                  "Note: 'Single' adults in this analysis refer to single-adult families, not single-adult households. In some cases, single adult families may share a household with other families. This differs from the analysis in the 'Gender' worksheet, where single adults are those who share the household with no other adults.",
                  "Note: Shared households such as such as a group of students, or other unrelated adults were excluded from the analysis.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Proportion of adults in each household food security category, Scotland 2019/20-2021/22",
                  "Number of adults in each household food security category, Scotland 2019/20-2021/22",
                  "Number of families in each group in the combined two-year survey sample, Scotland 2019/20-2021/22")),

  list(header = c("Household food security by marital status",
                  "This worksheet contains three tables.",
                  "Note: 'Single' refers to adults who have never been married or in a civil partnership, and are not living with a partner. 'Separated' refers to adults who are married or in a civil partnership, but are not living together because of estrangement. 'Married / Civil Partnership' includes couples who are temporarily living apart (e.g. due to serving in the armed forces). 'Adults' includes working-age adults and pensioners.",
                  "Note: Shared households such as such as a group of students, or other unrelated adults were excluded from the analysis.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Proportion of adults in each household food security category, Scotland 2019/20-2021/22",
                  "Number of adults in each household food security category, Scotland 2019/20-2021/22",
                  "Number of adults in each group in the combined two-year survey sample, Scotland 2019/20-2021/22")),

  # * -- income ----------------------------------------------------------------

  list(header = c("Median household income",
                  "This worksheet contains three tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c(paste0("Before housing costs: Median weekly equivalised household income in £ (in ", latestyear, " prices), Scotland"),
                  paste0("After housing costs: Median weekly equivalised household income in £ (in ", latestyear, " prices), Scotland"),
                  "Sample size: Number of families in each group in the combined three-year survey sample, Scotland")),
  list(header = c("Household income decile points",
                  "This worksheet contains three tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c(paste0("Before housing costs: Median weekly equivalised household income decile points in £ (in ", latestyear, " prices), Scotland"),
                  paste0("After housing costs: Median weekly equivalised household income decile points in £ (in ", latestyear, " prices), Scotland"),
                 "Sample size: Number of families in the combined three-year survey sample, Scotland")),
  list(header = c("Income decile shares",
                  "This worksheet contains three tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c(paste0("Before housing costs: Annual household income shares in £ million (in ", latestyear, " prices), Scotland"),
                  paste0("After housing costs: Annual household income shares in £ million (in ", latestyear, " prices), Scotland"),
                  "Sample size: Number of families in the combined three-year survey sample, Scotland")),
  list(header = c("Income inequality measures",
                  "This worksheet contains three tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c("Palma ratio (income share of the top 10% divided by the the bottom 40% of the household population), Scotland",
                  "Gini coefficient",
                  "Sample size: Number of families in the combined three-year survey sample, Scotland")),
  list(header = c("Poverty and other income thresholds",
                  "This worksheet contains four tables.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = c(paste0("Before housing costs: Weekly income in £, and after tax and transfers, Scotland ", latestperiod),
                  paste0("After housing costs: Annual income in £, and after tax and transfers, Scotland ", latestperiod),
                  paste0("Before housing costs: Weekly income in £, and after tax and transfers, Scotland ", latestperiod),
                  paste0("After housing costs: Annual income in £, after tax and transfers, Scotland ", latestperiod))),
  list(header = c("Income sources by income decile",
                  "This worksheet contains one table.",
                  "Important: The latest two estimates are two-year averages. See README for more information.",
                  "Source: Scottish Government analysis of the Family Resources Survey"),
       titles = paste0("Gross household income by income type as a share of total income, Scotland ",latestperiod))
  )

# add TOC link to each worksheet
link_to_toc <- '=HYPERLINK("#Contents!A1", "Table of contents")'

for (i in 1:length(myheaders)) {
  myheaders[[i]]$header <- append(myheaders[[i]]$header, link_to_toc, after = 1)
}

myheaders[[1]]$header

# number headers and tables ----------------------------------------------------
sheet_n <- length(mytables)
names(myheaders) <- paste0("sheet_", seq(1, sheet_n, 1))
names(mytables) <- paste0("sheet_", seq(1, sheet_n, 1))

# add prefixes to headings to mark them up as headings for VBA post-processing (optional)
for (i in 1:sheet_n) {
  myheaders[[i]]$header[1] <- paste0("H1-", myheaders[[i]]$header[1])
  myheaders[[i]]$titles <- paste0("H2-", myheaders[[i]]$titles)
}

# run all worksheets -----------------------------------------------------------

# Open / create workbook
if (file.exists(filename)) {wb <- loadWorkbook(filename)
} else {wb <- createWorkbook()}

invisible(
  sapply(seq(1, sheet_n, 1),
         function(x) {
           create_worksheet(sheet_no = x,
                            wb = wb,
                            tables = mytables[[x]],
                            headers = myheaders[[paste0("sheet_", x)]]$header,
                            titles = myheaders[[paste0("sheet_", x)]]$titles)
         }
  )
)

# create TOC -------------------------------------------------------------------
createContentSheet(wb = wb,
                   title = "H1-Table of contents",
                   toptext = "This workbook contains data used in the charts and tables in the Poverty and Income Inequality in Scotland National Statistics report, published in March 2023, and additional analysis. The report is available here: data.gov.scot/poverty",
                   headings = list(titles = c("H2-Headline poverty measures",
                                              "H2-People - relative and severe poverty after housing costs by a range of characteristics",
                                              "H2-Children - relative and severe poverty after housing costs by a range of characteristics",
                                              "H2-Adults - relative and severe poverty after housing costs by a range of characteristics",
                                              "H2-Household food security",
                                              "H2-Income and income inequality"),
                                   location = c(0, 8, 16, 30, 38, 49)))

# create Readme sheet ----------------------------------------------------------
createReadmeSheet(

  wb = wb,

  notes = list(
    "H1-Important notes" = c("Published: 23 March 2023",
                             "Next update: March 2024",
                             "The tables in this spreadsheet contain all estimates shown in the 'Poverty and Income Inequality in Scotland' National Statistics report, and further breakdowns. The report can be found here: https://data.gov.scot/poverty",
                             "Estimates are based on Scotland data from the Family Resources Survey. This survey is managed by the Department for Work and Pensions, who also published a report, available on their website.",
                             "Detailed information on definitions and methodology can be found in the report.",
                             "Estimates provided here are in most cases averages of three years of data. Three-year averages show trends better than single-year estimates. Single-year estimates of the headline measures are also available for download. Ethnicity and religion breakdowns are published as five-year averages."),

    "H2-COVID-19 impact" = c("Estimates for periods which include the year 2020/21 (the three-year periods 2018-21 and 2019-22, and the five-year periods 2016-21 and 2017-22) do not include the data collected during 2020/21. This means that tables with three-year averages contain two-year averages spanning three years for the latest two periods. Also, tables with five-year averages contain four-year averages spanning five years for the latest two periods.",
                             "The 2020/21 estimates are unreliable as they are based on data collected during the first year of the coronavirus (COVID-19) pandemic. Lockdown rules severely disrupted the data collection. As a result, we were unable to obtain a representative sample for Scotland."),

    "H2-Tables" = c("Each worksheet contains one or more tables. The main table is at the top, and extra tables are below. The extra tables give context such as sample sizes, average ages (for ethnicity and religion breakdowns), compositions, or population estimates.",
                    "Poverty rate tables show the proportion of individuals who are in poverty. Poverty rates describe the risk of being in poverty. If one group has a higher poverty rate than another, people in this group have a higher risk of being in poverty.",
                    "Composition tables show how people in poverty break down into subgroups. For example, it shows how many of those who are in poverty rent their homes versus how many own their homes. The poverty composition describes the scale of poverty for certain groups of people.",
                    "Numbers tables show the numbers of people (or children or adults) in poverty."),

    "H2-Poverty measures" = c("In this spreadsheet, we use different measures of poverty. Each measures a different aspect of people's living standards.",
                              "Relative poverty is the measure of low income that we use most often. It measures whether low-income households are keeping pace with middle income households. The relative poverty line is 60% of median UK income, and it changes every year along with median income.",
                              "Severe poverty is like relative poverty, but with a stricter threshold. It measures whether the lowest-income households are keeping pace with middle income households. The severe poverty line is 50% of median UK income, and it also changes every year.",
                              "Absolute poverty is another measure of low income. It measures whether the incomes of low-income households are keeping pace with inflation. The absolute poverty line is fixed at 60% of median UK income from 2010/11. It changes only a little due to inflation. It is now lower than the relative poverty line, but higher than the severe poverty line.",
                              "Material deprivation measures look at whether households can access basic goods and activities. The child measure combines material deprivation and low income. It looks at families who would like certain goods or activities but cannot afford them. The pensioner measure looks at pensioners who would like certain goods or activities but cannot access them. The barriers for this could be money, poor health, disability and social isolation."),

    "H2-Food security" = c("Food security questions were newly added to the Family Resources Survey in 2019/20. The household food security measure looks at how many people worry about running out of food or have to skip meals.",
                           "Note that we measure food security on a household level, but not everyone in the household always experiences this in the same way. For example, a parent may reduce their own meal sizes but not those of their young children. Note also that shared households such as such as a group of students, or other unrelated adults were excluded from the analysis."),

    "H2-Housing costs" = c("The most commonly used poverty indicator in Scotland is relative poverty after housing costs. After-housing-costs measures describe disposable household income after housing costs are paid for. They therefore better describe what is available to spend on food, bills and leisure every week or month.",
                           "In this spreadsheet, the headline poverty measures are available before and after housing costs. Poverty measures broken down by household and individual characteristics are available for poverty after housing costs. Estimates before housing costs are available at https://statistics.gov.scot."),

    "H2-Missing data" = c("In the tables, the following conventions have been used where figures are unavailable:",
                          "[u] - unreliable due to small sample size",
                          "[low] - below 0.5%",
                          "[b] - break in the time series",
                          "[x] - not available for another reason (data accuracy, data wasn't collected etc.)"),

    "H2-Reliability of the estimates" = c("The figures in these tables are estimates only, based on data from a sample survey. Thus, they could be a little bit higher or lower if we interviewed a different sample of the population.",
                                          "The precision of poverty rate estimates for the whole Scottish population is usually the central estimate plus or minus two percentage points, or plus or minus four percentage points for the Scottish child population. (Note that the precision is lower for 2018-21 estimates.) Precision is lower for smaller groups.",
                                          "Any small changes from year to year may not reflect real changes in the population. Longer-term trends are a better sign of a real change. Small differences between groups do not always reflect real differences in the population. Differences are more likely to be real if they are consistent over time.",
                                          "Estimates are considered unreliable when they are based on a very small sample size. For averages such as median incomes, this is less than 50 cases. For proportions and population estimates, it is less than 100 cases."),

    "H2-Rounding" = c("Proportions are rounded to five decimal places. Population numbers are rounded to the nearest 10,000 people. Annual GBP amounts are rounded to the nearest £100. Weekly GBP amounts are rounded to the nearest £1."),

    "H2-Data revision" = c("In 2021, previously published datasets underwent a minor methodological revision to capture all income from child maintenance. This led to small changes in household income and small adjustments to some poverty estimates. Therefore, some poverty and income estimates that have been published since March 2021 may not exactly match previously published estimates for 1994/95 to 2019/20. The revision did not affect any trends in poverty or household income."),

    "H2-Contact" = c("Maike Waldmann",
                     "Scottish Government",
                     "Communities Analysis Division",
                     "Email: social-justice-analysis@gov.scot")))

# save -------------------------------------------------------------------------

saveWorkbook(wb, filename, overwrite = TRUE)
cat(filename, " created", fill = TRUE)
rm(list = ls())
